# Use low memory map for in-memory index - about 45% less memory for index