@greatroar 2020-06-28
Interesting.

For comparison, I've been trying custom hash tables as well, but started with a simple chaning one with the duplicates in separate links. That also seemed to solve the memory use issue, but may come with some subtle performance issues: sequential performance was down between 8% and 31% on the benchmarks, and I could see a second core running the GC at 100%, so concurrent performance might be terrible.