@greatroar 2020-06-28
Also, I don't really get your 5 vs. 32 bytes comment, care to elaborate?