@aawsome 2020-06-28
> Do you know what restic command is most likely to allocate indexes and not much else?

I would say `restic cat blob` is your candidate :wink: 

> You're right about the duplicates handling, that's the part I couldn't quite get right.

IMO you should just include them into your chained list (aka linked list) and implement a `get` and `getAll` method like I did in this PR. This will save you 24 bytes per index entry.