@aawsome 2020-06-28
> Also, I don't really get your 5 vs. 32 bytes comment, care to elaborate?

The entry size for this PR is 32 bytes  + 16 bytes = 48 bytes. (49 bytes for your approach as you also save the BlobType)

For this PR: on average 1/2 byte of overhead per index entry + 1/2 unused page per 48 total pages. 1/2 page is 16*48 bytes and 48 pages will have around 32*48 entries. This is another 2 bytes overhead.

So I actually expect 2,5 bytes overhead per entry but worst case will be about double that amount.

For your approach I just took into account the extra pointer (8 bytes) and the duplicates slice (24 bytes). This is already 32 bytes per entry. On addition there is the overhead of managing the buckets which is 8 bytes per bucket and on average 2/3 buckets per entry. So total overhead should be 37 bytes on average. If you remove the duplicates slice, it will still be 13 bytes per entry on average and a bit more in the worst case.