@greatroar 2020-06-28
Judging from the [malloc code](https://github.com/golang/go/blob/c875503cf79f82b7d2dbec85667f330f72ec4557/src/runtime/sizeclasses.go), you're right about the tiny sizes. Good to know.